Use navigation from compose
