+++ app/build.gradle
@@ -43,7 +43,7 @@ android {
         compose true
     }
     composeOptions {
-        kotlinCompilerExtensionVersion composeVersion
+        kotlinCompilerExtensionVersion '1.2.0'
     }
     packagingOptions {
         resources {
@@ -63,6 +63,7 @@ dependencies {
 
     // Compose
     implementation "androidx.activity:activity-compose:$rootProject.activityComposeVersion"
+    implementation "androidx.navigation:navigation-compose:$rootProject.navigationComposeVersion"
     implementation "androidx.compose.ui:ui:$rootProject.composeVersion"
     implementation "androidx.compose.material:material:$rootProject.composeVersion"
     implementation "androidx.compose.ui:ui-tooling-preview:$rootProject.composeVersion"
